+++ CHANGELOG.md
@@ -42,6 +42,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### State Machine Breaking
 
 * [#224](https://github.com/babylonlabs-io/babylon/pull/224) Make injected checkpoint a standard tx
+* [#216](https://github.com/babylonlabs-io/babylon/pull/216) move voting power distribution
+update algorithm to `x/finality`
 * [#207](https://github.com/babylonlabs-io/babylon/pull/207) Rename total voting power
 to total bonded sat
 * [#204](https://github.com/babylonlabs-io/babylon/pull/204) Add babylon finality